+++ src/test/polymer-bundler_test.ts
@@ -24,16 +24,37 @@ const assert = chai.assert;
 
 suite('polymer-bundler CLI', () => {
 
+  const cliPath = path.resolve(__dirname, '../bin/polymer-bundler.js');
+
   test('uses the current working folder as loader root', async () => {
     const projectRoot = path.resolve(__dirname, '../../test/html');
-    const cli = path.resolve(__dirname, '../bin/polymer-bundler.js');
-    const stdout =
-        execSync([
-          `cd ${projectRoot}`,
-          `node ${cli} absolute-paths.html --inline-scripts --inline-css`,
-        ].join(' && '))
-            .toString();
+    const stdout = execSync([
+                     `cd ${projectRoot}`,
+                     `node ${cliPath} absolute-paths.html`,
+                   ].join(' && '))
+                       .toString();
     assert.include(stdout, '.absolute-paths-style');
     assert.include(stdout, 'hello from /absolute-paths/script.js');
+  });
+
+  suite('--redirect', () => {
+
+    test('handles urls with arbitrary protocols and hosts', async () => {
+      const projectRoot =
+          path.resolve(__dirname, '../../test/html/url-redirection')
+              // Force forward-slashes so quoting works with Windows paths.
+              .replace(/\\/g, '/');
+      const stdout =
+          execSync([
+            `cd ${projectRoot}`,
+            `node ${cliPath} index.html ` +
+                `--redirect="myapp://app/|${projectRoot}" ` +
+                `--redirect="vendor://|${projectRoot}/../bower_components"`,
+          ].join(' && '))
+              .toString();
+      assert.include(stdout, 'This is an external dependency');
+      assert.include(stdout, 'id="home-page"');
+      assert.include(stdout, 'id="settings-page"');
+    });
   });
 });