+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+- Added a `--redirect` option to `bin/polymer-bundler`.
 <!-- Add new, unreleased changes here. -->
 
 ## 2.0.3 - 2017-06-07